Create shadowsocks/client
