+++ client/client.go
@@ -1,28 +1,32 @@
+// Copyright 2023 Jigsaw Operations LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package client
 
 import (
-	"errors"
-	"io"
+	"fmt"
 	"net"
-	"time"
 
-	"github.com/Jigsaw-Code/outline-ss-server/internal/slicepool"
 	onet "github.com/Jigsaw-Code/outline-ss-server/net"
 	ss "github.com/Jigsaw-Code/outline-ss-server/shadowsocks"
-	"github.com/shadowsocks/go-shadowsocks2/socks"
+	ss_client "github.com/Jigsaw-Code/outline-ss-server/shadowsocks/client"
 )
 
-// clientUDPBufferSize is the maximum supported UDP packet size in bytes.
-const clientUDPBufferSize = 16 * 1024
-
-// udpPool stores the byte slices used for storing encrypted packets.
-var udpPool = slicepool.MakePool(clientUDPBufferSize)
-
 // Client is a client for Shadowsocks TCP and UDP connections.
+//
+// Deprecated: Use ss_client.StreamDialer and ss_client.PacketListener instead.
 type Client interface {
-	onet.StreamDialer
-	onet.PacketDialer
-
 	// DialTCP connects to `raddr` over TCP though a Shadowsocks proxy.
 	// `laddr` is a local bind address, a local address is automatically chosen if nil.
 	// `raddr` has the form `host:port`, where `host` can be a domain name or IP address.
@@ -44,156 +48,45 @@ type Client interface {
 
 // NewClient creates a client that routes connections to a Shadowsocks proxy listening at
 // `host:port`, with authentication parameters `cipher` (AEAD) and `password`.
-// TODO: add a dialer argument to support proxy chaining and transport changes.
+//
+// Deprecated: Use ss_client.StreamDialer and ss_client.PacketListener instead.
 func NewClient(host string, port int, password, cipherName string) (Client, error) {
 	// TODO: consider using net.LookupIP to get a list of IPs, and add logic for optimal selection.
 	proxyIP, err := net.ResolveIPAddr("ip", host)
 	if err != nil {
-		return nil, errors.New("Failed to resolve proxy address")
+		return nil, fmt.Errorf("Failed to resolve proxy address: %w", err)
 	}
 	cipher, err := ss.NewCipher(cipherName, password)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("Failed to create Shadowsocks cipher: %w", err)
 	}
-	d := ssClient{proxyIP: proxyIP.IP, proxyPort: port, cipher: cipher}
+	streamDialer, err := ss_client.NewStreamDialer(proxyIP.String(), port, cipher)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to create StreamDialer: %w", err)
+	}
+	packetListener, err := ss_client.NewPacketListener(proxyIP.String(), port, cipher)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to create PacketListener: %w", err)
+	}
+	d := ssClient{StreamDialer: streamDialer, PacketListener: packetListener}
 	return &d, nil
 }
 
 type ssClient struct {
-	proxyIP   net.IP
-	proxyPort int
-	cipher    *ss.Cipher
-	salter    ss.SaltGenerator
+	ss_client.StreamDialer
+	onet.PacketListener
 }
 
 func (c *ssClient) SetTCPSaltGenerator(salter ss.SaltGenerator) {
-	c.salter = salter
-}
-
-// This code contains an optimization to send the initial client payload along with
-// the Shadowsocks handshake.  This saves one packet during connection, and also
-// reduces the distinctiveness of the connection pattern.
-//
-// Normally, the initial payload will be sent as soon as the socket is connected,
-// except for delays due to inter-process communication.  However, some protocols
-// expect the server to send data first, in which case there is no client payload.
-// We therefore use a short delay, longer than any reasonable IPC but shorter than
-// typical network latency.  (In an Android emulator, the 90th percentile delay
-// was ~1 ms.)  If no client payload is received by this time, we connect without it.
-const helloWait = 10 * time.Millisecond
-
-func (c *ssClient) Dial(addr string) (onet.DuplexConn, error) {
-	return c.DialTCP(nil, addr)
+	c.StreamDialer.SetTCPSaltGenerator(salter)
 }
 
 func (c *ssClient) DialTCP(laddr *net.TCPAddr, raddr string) (onet.DuplexConn, error) {
-	socksTargetAddr := socks.ParseAddr(raddr)
-	if socksTargetAddr == nil {
-		return nil, errors.New("Failed to parse target address")
-	}
-	proxyAddr := &net.TCPAddr{IP: c.proxyIP, Port: c.proxyPort}
-	proxyConn, err := net.DialTCP("tcp", laddr, proxyAddr)
-	if err != nil {
-		return nil, err
-	}
-	ssw := ss.NewShadowsocksWriter(proxyConn, c.cipher)
-	if c.salter != nil {
-		ssw.SetSaltGenerator(c.salter)
-	}
-	_, err = ssw.LazyWrite(socksTargetAddr)
-	if err != nil {
-		proxyConn.Close()
-		return nil, errors.New("Failed to write target address")
-	}
-	time.AfterFunc(helloWait, func() {
-		ssw.Flush()
-	})
-	ssr := ss.NewShadowsocksReader(proxyConn, c.cipher)
-	return onet.WrapConn(proxyConn, ssr, ssw), nil
-}
-
-func (c *ssClient) ListenPacket() (net.PacketConn, error) {
-	return c.ListenUDP(nil)
+	// TODO: restore laddr support
+	return c.Dial(raddr)
 }
 
 func (c *ssClient) ListenUDP(laddr *net.UDPAddr) (net.PacketConn, error) {
-	proxyAddr := &net.UDPAddr{IP: c.proxyIP, Port: c.proxyPort}
-	pc, err := net.DialUDP("udp", laddr, proxyAddr)
-	if err != nil {
-		return nil, err
-	}
-	conn := packetConn{UDPConn: pc, cipher: c.cipher}
-	return &conn, nil
-}
-
-type packetConn struct {
-	*net.UDPConn
-	cipher *ss.Cipher
-}
-
-// WriteTo encrypts `b` and writes to `addr` through the proxy.
-func (c *packetConn) WriteTo(b []byte, addr net.Addr) (int, error) {
-	socksTargetAddr := socks.ParseAddr(addr.String())
-	if socksTargetAddr == nil {
-		return 0, errors.New("Failed to parse target address")
-	}
-	lazySlice := udpPool.LazySlice()
-	cipherBuf := lazySlice.Acquire()
-	defer lazySlice.Release()
-	saltSize := c.cipher.SaltSize()
-	// Copy the SOCKS target address and payload, reserving space for the generated salt to avoid
-	// partially overlapping the plaintext and cipher slices since `Pack` skips the salt when calling
-	// `AEAD.Seal` (see https://golang.org/pkg/crypto/cipher/#AEAD).
-	plaintextBuf := append(append(cipherBuf[saltSize:saltSize], socksTargetAddr...), b...)
-	buf, err := ss.Pack(cipherBuf, plaintextBuf, c.cipher)
-	if err != nil {
-		return 0, err
-	}
-	_, err = c.UDPConn.Write(buf)
-	return len(b), err
-}
-
-// ReadFrom reads from the embedded PacketConn and decrypts into `b`.
-func (c *packetConn) ReadFrom(b []byte) (int, net.Addr, error) {
-	lazySlice := udpPool.LazySlice()
-	cipherBuf := lazySlice.Acquire()
-	defer lazySlice.Release()
-	n, err := c.UDPConn.Read(cipherBuf)
-	if err != nil {
-		return 0, nil, err
-	}
-	// Decrypt in-place.
-	buf, err := ss.Unpack(nil, cipherBuf[:n], c.cipher)
-	if err != nil {
-		return 0, nil, err
-	}
-	socksSrcAddr := socks.SplitAddr(buf)
-	if socksSrcAddr == nil {
-		return 0, nil, errors.New("Failed to read source address")
-	}
-	srcAddr := newAddr(socksSrcAddr.String(), "udp")
-	n = copy(b, buf[len(socksSrcAddr):]) // Strip the SOCKS source address
-	if len(b) < len(buf)-len(socksSrcAddr) {
-		return n, srcAddr, io.ErrShortBuffer
-	}
-	return n, srcAddr, nil
-}
-
-type addr struct {
-	address string
-	network string
-}
-
-func (a *addr) String() string {
-	return a.address
-}
-
-func (a *addr) Network() string {
-	return a.network
-}
-
-// newAddr returns a net.Addr that holds an address of the form `host:port` with a domain name or IP as host.
-// Used for SOCKS addressing.
-func newAddr(address, network string) net.Addr {
-	return &addr{address: address, network: network}
+	// TODO: restore laddr support
+	return c.ListenPacket()
 }
