+++ shadowsocks/client/salt_test.go
