+++ shadowsocks/client/stream_dialer_test.go
@@ -1,0 +1,242 @@
+// Copyright 2023 Jigsaw Operations LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package client
+
+import (
+	"io"
+	"net"
+	"net/netip"
+	"sync"
+	"testing"
+	"time"
+
+	onet "github.com/Jigsaw-Code/outline-ss-server/net"
+	ss "github.com/Jigsaw-Code/outline-ss-server/shadowsocks"
+	"github.com/shadowsocks/go-shadowsocks2/socks"
+)
+
+func TestShadowsocksClient_DialTCP(t *testing.T) {
+	cipher, err := ss.NewCipher(ss.TestCipher, testPassword)
+	if err != nil {
+		t.Fatalf("Failed to create cipher: %v", err)
+	}
+	proxy, running := startShadowsocksTCPEchoProxy(cipher, testTargetAddr, t)
+	proxyAddr := netip.MustParseAddrPort(proxy.Addr().String())
+	d, err := NewStreamDialer(proxyAddr.Addr().String(), int(proxyAddr.Port()), cipher)
+	if err != nil {
+		t.Fatalf("Failed to create ShadowsocksClient: %v", err)
+	}
+	conn, err := d.Dial(testTargetAddr)
+	if err != nil {
+		t.Fatalf("ShadowsocksClient.DialTCP failed: %v", err)
+	}
+	conn.SetReadDeadline(time.Now().Add(time.Second * 5))
+	expectEchoPayload(conn, ss.MakeTestPayload(1024), make([]byte, 1024), t)
+	conn.Close()
+
+	proxy.Close()
+	running.Wait()
+}
+
+func TestShadowsocksClient_DialTCPNoPayload(t *testing.T) {
+	cipher, err := ss.NewCipher(ss.TestCipher, testPassword)
+	if err != nil {
+		t.Fatalf("Failed to create cipher: %v", err)
+	}
+	proxy, running := startShadowsocksTCPEchoProxy(cipher, testTargetAddr, t)
+	proxyAddr := netip.MustParseAddrPort(proxy.Addr().String())
+	d, err := NewStreamDialer(proxyAddr.Addr().String(), int(proxyAddr.Port()), cipher)
+	if err != nil {
+		t.Fatalf("Failed to create ShadowsocksClient: %v", err)
+	}
+	conn, err := d.Dial(testTargetAddr)
+	if err != nil {
+		t.Fatalf("ShadowsocksClient.DialTCP failed: %v", err)
+	}
+
+	// Wait for more than 10 milliseconds to ensure that the target
+	// address is sent.
+	time.Sleep(20 * time.Millisecond)
+	// Force the echo server to verify the target address.
+	conn.Close()
+
+	proxy.Close()
+	running.Wait()
+}
+
+func TestShadowsocksClient_DialTCPFastClose(t *testing.T) {
+	// Set up a listener that verifies no data is sent.
+	listener, err := net.ListenTCP("tcp", &net.TCPAddr{IP: net.ParseIP("127.0.0.1"), Port: 0})
+	if err != nil {
+		t.Fatalf("ListenTCP failed: %v", err)
+	}
+
+	done := make(chan struct{})
+	go func() {
+		conn, err := listener.Accept()
+		if err != nil {
+			t.Error(err)
+		}
+		buf := make([]byte, 64)
+		n, err := conn.Read(buf)
+		if n > 0 || err != io.EOF {
+			t.Errorf("Expected EOF, got %v, %v", buf[:n], err)
+		}
+		listener.Close()
+		close(done)
+	}()
+
+	cipher, err := ss.NewCipher(ss.TestCipher, testPassword)
+	if err != nil {
+		t.Fatalf("Failed to create cipher: %v", err)
+	}
+	proxyAddr := netip.MustParseAddrPort(listener.Addr().String())
+	d, err := NewStreamDialer(proxyAddr.Addr().String(), int(proxyAddr.Port()), cipher)
+	if err != nil {
+		t.Fatalf("Failed to create ShadowsocksClient: %v", err)
+	}
+
+	conn, err := d.Dial(testTargetAddr)
+	if err != nil {
+		t.Fatalf("ShadowsocksClient.DialTCP failed: %v", err)
+	}
+
+	// Wait for less than 10 milliseconds to ensure that the target
+	// address is not sent.
+	time.Sleep(1 * time.Millisecond)
+	// Close the connection before the target address is sent.
+	conn.Close()
+	// Wait for the listener to verify the close.
+	<-done
+}
+
+func TestShadowsocksClient_TCPPrefix(t *testing.T) {
+	prefix := []byte("test prefix")
+
+	listener, err := net.ListenTCP("tcp", &net.TCPAddr{IP: net.ParseIP("127.0.0.1"), Port: 0})
+	if err != nil {
+		t.Fatalf("ListenTCP failed: %v", err)
+	}
+	var running sync.WaitGroup
+	running.Add(1)
+	go func() {
+		defer running.Done()
+		defer listener.Close()
+		clientConn, err := listener.AcceptTCP()
+		if err != nil {
+			t.Logf("AcceptTCP failed: %v", err)
+			return
+		}
+		defer clientConn.Close()
+		prefixReceived := make([]byte, len(prefix))
+		if _, err := io.ReadFull(clientConn, prefixReceived); err != nil {
+			t.Error(err)
+		}
+		for i := range prefix {
+			if prefixReceived[i] != prefix[i] {
+				t.Error("prefix contents mismatch")
+			}
+		}
+	}()
+
+	cipher, err := ss.NewCipher(ss.TestCipher, testPassword)
+	if err != nil {
+		t.Fatalf("Failed to create cipher: %v", err)
+	}
+	proxyAddr := netip.MustParseAddrPort(listener.Addr().String())
+	d, err := NewStreamDialer(proxyAddr.Addr().String(), int(proxyAddr.Port()), cipher)
+	if err != nil {
+		t.Error(err)
+	}
+	d.SetTCPSaltGenerator(NewPrefixSaltGenerator(prefix))
+	conn, err := d.Dial(testTargetAddr)
+	if err != nil {
+		t.Fatalf("ShadowsocksClient.DialTCP failed: %v", err)
+	}
+	conn.Write(nil)
+	conn.Close()
+	running.Wait()
+}
+
+func BenchmarkShadowsocksClient_DialTCP(b *testing.B) {
+	b.StopTimer()
+	b.ResetTimer()
+
+	cipher, err := ss.NewCipher(ss.TestCipher, testPassword)
+	if err != nil {
+		b.Fatalf("Failed to create cipher: %v", err)
+	}
+	proxy, running := startShadowsocksTCPEchoProxy(cipher, testTargetAddr, b)
+	proxyAddr := netip.MustParseAddrPort(proxy.Addr().String())
+	d, err := NewStreamDialer(proxyAddr.Addr().String(), int(proxyAddr.Port()), cipher)
+	if err != nil {
+		b.Fatalf("Failed to create ShadowsocksClient: %v", err)
+	}
+	conn, err := d.Dial(testTargetAddr)
+	if err != nil {
+		b.Fatalf("ShadowsocksClient.DialTCP failed: %v", err)
+	}
+	conn.SetReadDeadline(time.Now().Add(time.Second * 5))
+	buf := make([]byte, 1024)
+	for n := 0; n < b.N; n++ {
+		payload := ss.MakeTestPayload(1024)
+		b.StartTimer()
+		expectEchoPayload(conn, payload, buf, b)
+		b.StopTimer()
+	}
+
+	conn.Close()
+	proxy.Close()
+	running.Wait()
+}
+
+func startShadowsocksTCPEchoProxy(cipher *ss.Cipher, expectedTgtAddr string, t testing.TB) (net.Listener, *sync.WaitGroup) {
+	listener, err := net.ListenTCP("tcp", &net.TCPAddr{IP: net.ParseIP("127.0.0.1"), Port: 0})
+	if err != nil {
+		t.Fatalf("ListenTCP failed: %v", err)
+	}
+	t.Logf("Starting SS TCP echo proxy at %v\n", listener.Addr())
+	var running sync.WaitGroup
+	running.Add(1)
+	go func() {
+		defer running.Done()
+		defer listener.Close()
+		for {
+			clientConn, err := listener.AcceptTCP()
+			if err != nil {
+				t.Logf("AcceptTCP failed: %v", err)
+				return
+			}
+			running.Add(1)
+			go func() {
+				defer running.Done()
+				defer clientConn.Close()
+				ssr := ss.NewShadowsocksReader(clientConn, cipher)
+				ssw := ss.NewShadowsocksWriter(clientConn, cipher)
+				ssClientConn := onet.WrapConn(clientConn, ssr, ssw)
+
+				tgtAddr, err := socks.ReadAddr(ssClientConn)
+				if err != nil {
+					t.Fatalf("Failed to read target address: %v", err)
+				}
+				if tgtAddr.String() != expectedTgtAddr {
+					t.Fatalf("Expected target address '%v'. Got '%v'", expectedTgtAddr, tgtAddr)
+				}
+				io.Copy(ssw, ssr)
+			}()
+		}
+	}()
+	return listener, &running
+}
